feat(ui): T5.3 + T5.4 Android bottom-tab nav, reduced-motion, visual QA
T5.3 — Android bottom-tab navigation:
- New `components/bottom_tab_bar.slint`: 56px 4-tab bar with accent
  gradient indicator for active tab; uses Theme tokens throughout.
- `MainWindow.slint`: add `in property <length> screen-width` (default
  900px = desktop sidebar layout). When < 600px the sidebar is replaced
  by `BottomTabBar` at the bottom (phone/Android layout). When ≥ 600px
  the existing `HorizontalLayout { Sidebar; content; DetailPanel }` is
  rendered unchanged.
- `root_window.rs`: call `window.set_screen_width(900.0)` at startup
  (desktop default); Android init path overrides this with the physical
  device width / scale factor.

T5.4 — Motion, reduced-motion, visual QA:
- `theme.slint`: add `reduced-motion` bool to `Theme` global + six
  computed `anim-*` duration tokens (sidebar 180ms, panel 200ms,
  row 80ms, palette 140ms, toast 160ms, toggle 120ms). When
  `reduced-motion` is true all durations clamp to 0ms except toast
  (100ms minimum for accessibility visibility).
- Replaced all hard-coded `duration: Xms` across every `.slint` file
  with the appropriate `Theme.anim-*` token so reduced-motion works
  system-wide without touching each component individually.
- `command_palette.slint`: add `animate opacity { duration:
  Theme.anim-palette; easing: ease-out; }` for the 140ms appear fade.
- `root_window.rs`: call `NSWorkspace accessibilityDisplayShouldReduceMotion`
  on macOS at startup and forward the result to `Theme.set_reduced_motion`;
  hardcoded `false` on non-macOS platforms.
- `history_model.rs`: fix pre-existing clippy `unnecessary_cast` (u64 as u64).
- `daemon.rs`, `migration_gate_clears.rs`: rustfmt line-wrap normalisations
  (pre-existing fmt drift, no logic change).

Build: `cargo build -p copypaste-ui` clean.
Clippy: `cargo clippy -p copypaste-ui -- -D warnings` clean.
Fmt: `cargo fmt --all -- --check` clean.

+++ crates/copypaste-ui/src/history_model.rs
@@ -105,7 +105,7 @@ impl HistoryModel {
     /// `true` when the daemon has more items beyond what is already loaded.
     #[allow(dead_code)]
     fn has_more(&self) -> bool {
-        (self.next_offset.get() as u64) < self.total_count.get()
+        self.next_offset.get() < self.total_count.get()
             || self.total_count.get() == 0 && self.next_offset.get() == 0
     }
 
@@ -167,11 +167,10 @@ impl HistoryModel {
 
         // Build the IPC result; restore query + clear loading flag on all
         // paths (including the early-return from `connect` failure).
-        let result: anyhow::Result<crate::ipc_client::HistoryPage> =
-            match self.connect() {
-                None => Err(anyhow::anyhow!("daemon offline")),
-                Some(mut client) => client.history_page(self.page_size, offset),
-            };
+        let result: anyhow::Result<crate::ipc_client::HistoryPage> = match self.connect() {
+            None => Err(anyhow::anyhow!("daemon offline")),
+            Some(mut client) => client.history_page(self.page_size, offset),
+        };
 
         // Restore the query field and clear the loading flag regardless of
         // whether the IPC call succeeded or failed.
@@ -183,8 +182,7 @@ impl HistoryModel {
                 self.total_count.set(page.total);
                 let new_offset = offset + page.items.len() as u64;
                 self.next_offset.set(new_offset);
-                let clip_items: Vec<ClipItem> =
-                    page.items.iter().map(entry_to_clip_item).collect();
+                let clip_items: Vec<ClipItem> = page.items.iter().map(entry_to_clip_item).collect();
                 self.append_items(clip_items);
                 Ok(())
             }
